Divided code into files self. changes ETX at the end of the message added
